--- a/.gitignore
+++ b/.gitignore
@@ -29,3 +29,5 @@ venv/
 *.swo
 
 dev/
+
+*copy*