--- a/notebooks/0-example-with-warper.ipynb
+++ b/notebooks/0-example-with-warper.ipynb
@@ -94,7 +94,7 @@
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1239bfe10>"
+       "<pywarper.warper.Warper at 0x1135e9dd0>"
       ]
      },
      "execution_count": 4,
@@ -118,15 +118,15 @@
      "text": [
       "[Warper] Building mapping …\n",
       "Mapping min position (On SAC layer)...\n",
-      "Mapping min position completed in 3.80 seconds.\n",
+      "Mapping min position completed in 0.83 seconds.\n",
       "Mapping max position (Off SAC layer)...\n",
-      "Mapping max position completed in 3.61 seconds.\n"
+      "Mapping max position completed in 0.79 seconds.\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1239bfe10>"
+       "<pywarper.warper.Warper at 0x1135e9dd0>"
       ]
      },
      "execution_count": 5,
@@ -150,13 +150,13 @@
      "text": [
       "[Warper] Warping arbor …\n",
       "Warping nodes...\n",
-      "Nodes warped in 4.35 seconds.\n"
+      "Nodes warped in 1.23 seconds.\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1239bfe10>"
+       "<pywarper.warper.Warper at 0x1135e9dd0>"
       ]
      },
      "execution_count": 6,
@@ -171,14 +171,14 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 7,
    "id": "f6aa5afe",
    "metadata": {},
    "outputs": [
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1239bfe10>"
+       "<pywarper.warper.Warper at 0x1135e9dd0>"
       ]
      },
      "execution_count": 7,
@@ -282,7 +282,7 @@
     {
      "data": {
       "text/plain": [
-       "<matplotlib.legend.Legend at 0x124fcccd0>"
+       "<matplotlib.legend.Legend at 0x116127350>"
       ]
      },
      "execution_count": 10,
@@ -350,14 +350,14 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "Last updated: 2025-05-15 13:08:29CEST\n",
+      "Last updated: 2025-05-20 00:03:01CEST\n",
       "\n",
       "Python implementation: CPython\n",
       "Python version       : 3.11.11\n",
       "IPython version      : 9.1.0\n",
       "\n",
-      "numpy     : 2.2.4\n",
       "pandas    : 2.2.3\n",
+      "numpy     : 2.2.4\n",
       "matplotlib: 3.10.1\n",
       "pywarper  : 0.1.4\n",
       "\n",