--- a/notebooks/0-example-with-warper.ipynb
+++ b/notebooks/0-example-with-warper.ipynb
@@ -28,7 +28,7 @@
     "    ----------\n",
     "    fname\n",
     "        Plain‐text file with at least the columns *X*, *Y*, *Slice* as in the\n",
-    "        Sümbül *et al.* (2014) dataset.\n",
+    "        Sümbül *et al.* (2014) dataset.\n",
     "\n",
     "    Returns\n",
     "    -------\n",
@@ -94,7 +94,7 @@
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1135e9dd0>"
+       "<pywarper.warper.Warper at 0x1092eb5d0>"
       ]
      },
      "execution_count": 4,
@@ -118,15 +118,15 @@
      "text": [
       "[Warper] Building mapping …\n",
       "Mapping min position (On SAC layer)...\n",
-      "Mapping min position completed in 0.83 seconds.\n",
+      "Mapping min position completed in 0.88 seconds.\n",
       "Mapping max position (Off SAC layer)...\n",
-      "Mapping max position completed in 0.79 seconds.\n"
+      "Mapping max position completed in 0.84 seconds.\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1135e9dd0>"
+       "<pywarper.warper.Warper at 0x1092eb5d0>"
       ]
      },
      "execution_count": 5,
@@ -150,13 +150,13 @@
      "text": [
       "[Warper] Warping arbor …\n",
       "Warping nodes...\n",
-      "Nodes warped in 1.23 seconds.\n"
+      "Nodes warped in 1.35 seconds.\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1135e9dd0>"
+       "<pywarper.warper.Warper at 0x1092eb5d0>"
       ]
      },
      "execution_count": 6,
@@ -178,7 +178,7 @@
     {
      "data": {
       "text/plain": [
-       "<pywarper.warper.Warper at 0x1135e9dd0>"
+       "<pywarper.warper.Warper at 0x1092eb5d0>"
       ]
      },
      "execution_count": 7,
@@ -282,7 +282,7 @@
     {
      "data": {
       "text/plain": [
-       "<matplotlib.legend.Legend at 0x116127350>"
+       "<matplotlib.legend.Legend at 0x11ea282d0>"
       ]
      },
      "execution_count": 10,
@@ -350,16 +350,16 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "Last updated: 2025-05-20 00:03:01CEST\n",
+      "Last updated: 2025-05-20 09:27:43CEST\n",
       "\n",
       "Python implementation: CPython\n",
       "Python version       : 3.11.11\n",
       "IPython version      : 9.1.0\n",
       "\n",
-      "pandas    : 2.2.3\n",
-      "numpy     : 2.2.4\n",
       "matplotlib: 3.10.1\n",
-      "pywarper  : 0.1.4\n",
+      "numpy     : 2.2.4\n",
+      "pandas    : 2.2.3\n",
+      "pywarper  : 0.1.5\n",
       "\n",
       "Watermark: 2.5.0\n",
       "\n"