--- a/notebooks/1-example-step-by-step.ipynb
+++ b/notebooks/1-example-step-by-step.ipynb
@@ -74,9 +74,9 @@
      "output_type": "stream",
      "text": [
       "Mapping min position (On SAC layer)...\n",
-      "Mapping min position completed in 0.79 seconds.\n",
+      "Mapping min position completed in 0.88 seconds.\n",
       "Mapping max position (Off SAC layer)...\n",
-      "Mapping max position completed in 0.81 seconds.\n"
+      "Mapping max position completed in 1.09 seconds.\n"
      ]
     }
    ],
@@ -97,7 +97,7 @@
      "output_type": "stream",
      "text": [
       "Warping nodes...\n",
-      "Nodes warped in 1.17 seconds.\n"
+      "Nodes warped in 1.41 seconds.\n"
      ]
     }
    ],
@@ -306,7 +306,7 @@
     {
      "data": {
       "text/plain": [
-       "<matplotlib.legend.Legend at 0x11e346f10>"
+       "<matplotlib.legend.Legend at 0x1224aaed0>"
       ]
      },
      "execution_count": 11,
@@ -374,17 +374,17 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "Last updated: 2025-05-20 00:00:38CEST\n",
+      "Last updated: 2025-05-20 09:28:10CEST\n",
       "\n",
       "Python implementation: CPython\n",
       "Python version       : 3.11.11\n",
       "IPython version      : 9.1.0\n",
       "\n",
-      "pywarper  : 0.1.4\n",
-      "scipy     : 1.15.2\n",
+      "pandas    : 2.2.3\n",
       "numpy     : 2.2.4\n",
       "matplotlib: 3.10.1\n",
-      "pandas    : 2.2.3\n",
+      "scipy     : 1.15.2\n",
+      "pywarper  : 0.1.5\n",
       "\n",
       "Watermark: 2.5.0\n",
       "\n"