--- a/pyproject.toml
+++ b/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "pywarper"
-version = "0.1.4"
+version = "0.1.5"
 description = "Conformal mapping-based warping of neuronal arbor morphologies."
 authors = []
 requires-python = ">=3.9.0"
@@ -22,8 +22,9 @@ readme = {file = "README.md", content-type = "text/markdown"}
 dev = [
     "ruff",
     "pytest",
-    "twine", 
+    "twine",
     "maturin",
+    "ipykernel>=6.29.5",
 ]
 scikit-sparse = [
     "scikit-sparse>=0.4.15",
@@ -33,4 +34,3 @@ scikit-sparse = [
 [tool.setuptools]
 packages = {find = {}}
 include-package-data = true
-