```bash
hyperfine \      
  --command-name HEAD \
  --prepare 'git checkout -q HEAD' \
  'pytest tests/test_arbor.py -q --cache-clear' \
  --command-name HEAD~1 \
  --prepare 'git checkout -q HEAD~1' \
  'pytest tests/test_arbor.py -q --cache-clear'

Benchmark 1: HEAD
  Time (mean ± σ):      5.524 s ±  0.515 s    [User: 10.947 s, System: 1.846 s]
  Range (min … max):    4.999 s …  6.195 s    10 runs
 
Benchmark 2: HEAD~1
  Time (mean ± σ):     13.858 s ±  1.669 s    [User: 17.881 s, System: 1.831 s]
  Range (min … max):   12.383 s … 17.099 s    10 runs
 
Summary
  HEAD ran
    2.51 ± 0.38 times faster than HEAD~1
```